fix: extract trace level enum in a separate header file
