+++ extension-host/include/sqlextensionhostcallbacks.h
@@ -1,7 +1,7 @@
 //*********************************************************************
 //      Copyright (c) Microsoft Corporation.
 //
-// Host callbacks protocol between Exthost and a 3rd party extension.
+// Host callbacks protocol between the host and extensions.
 // Optional API surface introduced alongside v3 of the Extension API.
 //
 // @File: sqlextensionhostcallbacks.h
@@ -12,23 +12,11 @@
 #define __SQLEXTENSIONHOSTCALLBACKS
 
 #include "sqlexternallanguage.h"
+#include "sqlextensiontracelevel.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif /* __cplusplus */
-
-// Trace levels for events logged via LogXEvent function.
-// Lowest numeric value is the most severe, matching the Windows ETW
-// TRACE_LEVEL_* convention.
-//
-enum ExtensionTraceLevel
-{
-	Extension_Critical    = 1,
-	Extension_Error       = 2,
-	Extension_Warning     = 3,
-	Extension_Information = 4,
-	Extension_Verbose     = 5
-};
 
 // Callback function provided by the host for logging XEvents from Extension.
 //
@@ -50,8 +38,30 @@ typedef void (*PFunc_ExtensionLogXEvent)(
 
 typedef struct _SQLEXTENSION_HOST_CALLBACKS
 {
+	// Highest SQLEXTENSION_HOST_CALLBACKS_VERSION_* the host populates.
+	// Extension must validate before reading version-gated fields.
+	//
 	SQLUSMALLINT Version;
+
+	// Explicit padding so SizeInBytes is naturally 4-byte aligned regardless
+	// of compiler packing settings. Must be zero.
+	//
+	SQLUSMALLINT Reserved0;
+
+	// sizeof(SQLEXTENSION_HOST_CALLBACKS) as the host saw it at build time.
+	// Extension must validate this is greater or equal the size of every field it intends
+	// to read. Lets a newer Extension safely run against an older host that supplied a smaller struct.
+	//
+	SQLUINTEGER SizeInBytes;
+
+	// Version 1 callbacks.
+	//
 	PFunc_ExtensionLogXEvent LogXEvent;
+
+	// Reserved for future expansion. Zero-initialized by the host. Extension must not read or call these.
+	//
+	void *Reserved1;
+	void *Reserved2;
 } SQLEXTENSION_HOST_CALLBACKS;
 
 // Optional API (v3+)
